Create your first Rust Serverless function
With our crate in place we can set up our lambda code for the Netlify Function. We'll need `tokio`, `serde_json`, and `lambda_runtime`.

```diff
cargo add -p pokemon-api tokio lambda_runtime serde_json
```

`tokio` is an async runtime and we need it because the function signature we'll be working with from `lambda_runtime` is async.

`lambda_runtime` is the Rust runtime for AWS Lambda maintained by a team at AWS, which is what Netlify runs on behind the scenes. It's responsible for the mechanics of how lambda requests get handled.

`serde_json` is used for both deserializing incoming data and serializing outgoing data from and to JSON.

```rust
use lambda_runtime::{handler_fn, Context, Error};
use serde_json::{json, Value};

async fn main() -> Result<(), Error> {
    println!("cold start");
    let processor = handler_fn(handler);
    lambda_runtime::run(processor).await?;
    Ok(())
}

async fn handler(
    _: Value,
    _: Context,
) -> Result<Value, Error> {
    println!("handler");
    Ok(json!({"body": "Boop!"}))
}
```

The setup for a lambda function can be through of as two phases:

- what happens on a cold start
- what happens to process a single incoming event

In our case, the `main` function runs on cold start, which means it runs once for each instance of our function Netlify or AWS would build up. Another instance of a function is generally started if there aren't enough instances available to handle the current requests.

The `lambda_runtime::run` call in `main` internally bootstraps a `Stream` while loop to process each event, just like we did in the Pokemon csv upload course, except this one never completes.

The `handler` function is the one that the `lambda_runtime::run` processing uses to process each event.

```rust
async fn main() -> Result<(), Error> {
    println!("cold start");
    let processor = handler_fn(handler);
    lambda_runtime::run(processor).await?;
    Ok(())
}
```

The main function is wrapped in the `tokio::main` macro, which wraps our application in the tokio async runtime. That's what lets us `await` the lambda runtime and process each event.

`println!` is being used so that later when we look at the logs we can easily see which pieces of our code are running at what times. This is the section that only runs during the cold start.

The lambda runtime crate includes a trait called `Handler`. Any struct that implements `Handler` can be passed into the `lambda_runtime::run` function. `handler_fn` creates a struct that implements this trait for us, it's a lot like the pattern of calling `::new` to construct a new struct that we used in the `PokemonId` from the csv upload course.

The `processor` can then be passed into the `lambda_runtime::run` function where it will be used to call our `handler` function for each event.

We `await` here, which for our thought models purposes never returns. It waits in a loop for new events to come in forever as long as the lambda is alive.

Our `handler` is much smaller and does less.

```rust
async fn handler(
    _: Value,
    _: Context,
) -> Result<Value, Error> {
    println!("handler");
    Ok(json!({"body": "Boop!"}))
}
```

The `handler` function accepts two arguments. The first is a `serde_json::Value`, which is a JSON blob that looks like this.

```json
{
  "body": "",
  "headers": {
    "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
    "accept-encoding": "br, gzip",
    "accept-language": "en-US,en;q=0.9,fr;q=0.8",
    "cache-control": "no-cache",
    "client-ip": "100.64.0.115",
    "connection": "keep-alive",
    "dnt": "1",
    "forwarded": "for=67.180.63.210;proto=https",
    "host": "netlify-rust-test-function.netlify.app",
    "pragma": "no-cache",
    "sec-ch-ua": "\"Chromium\";v=\"92\", \" Not A;Brand\";v=\"99\", \"Google Chrome\";v=\"92\"",
    "sec-ch-ua-mobile": "?0",
    "sec-fetch-dest": "document",
    "sec-fetch-mode": "navigate",
    "sec-fetch-site": "none",
    "sec-fetch-user": "?1",
    "upgrade-insecure-requests": "1",
    "user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.107 Safari/537.36",
    "via": "http/1.1 Netlify[6826df03-0183-43f0-986e-2b809e4bab67] (Netlify Edge Server)",
    "x-bb-ab": "0.541370",
    "x-bb-client-request-uuid": "01FFEGH1HV7GPWHK5HCGQ4AAAV",
    "x-bb-ip": "67.180.63.210",
    "x-bb-loop": "1",
    "x-cdn-domain": "www.bitballoon.com",
    "x-country": "US",
    "x-datadog-parent-id": "1524629822969833135",
    "x-datadog-sampling-priority": "1",
    "x-datadog-trace-id": "15496261542005402192",
    "x-forwarded-for": "67.180.63.210, 100.64.0.115",
    "x-forwarded-proto": "https",
    "x-language": "en,en;q=0.9,fr;q=0.8",
    "x-nf-client-connection-ip": "67.180.63.210",
    "x-nf-connection-proto": "https",
    "x-nf-request-id": "01FFEGH1HV7GPWHK5HCGQ4AAAV"
  },
  "httpMethod": "GET",
  "isBase64Encoded": false,
  "multiValueHeaders": {
    "Accept": [
      "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9"
    ],
    "Accept-Encoding": ["br, gzip"],
    "Accept-Language": ["en-US,en;q=0.9,fr;q=0.8"],
    "Cache-Control": ["no-cache"],
    "Client-Ip": ["100.64.0.115"],
    "Connection": ["keep-alive"],
    "Dnt": ["1"],
    "Forwarded": ["for=67.180.63.210;proto=https"],
    "Pragma": ["no-cache"],
    "Sec-Ch-Ua": [
      "\"Chromium\";v=\"92\", \" Not A;Brand\";v=\"99\", \"Google Chrome\";v=\"92\""
    ],
    "Sec-Ch-Ua-Mobile": ["?0"],
    "Sec-Fetch-Dest": ["document"],
    "Sec-Fetch-Mode": ["navigate"],
    "Sec-Fetch-Site": ["none"],
    "Sec-Fetch-User": ["?1"],
    "Upgrade-Insecure-Requests": ["1"],
    "User-Agent": [
      "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.107 Safari/537.36"
    ],
    "Via": [
      "http/1.1 Netlify[6826df03-0183-43f0-986e-2b809e4bab67] (Netlify Edge Server)"
    ],
    "X-Bb-Ab": ["0.541370"],
    "X-Bb-Client-Request-Uuid": ["01FFEGH1HV7GPWHK5HCGQ4AAAV"],
    "X-Bb-Ip": ["67.180.63.210"],
    "X-Bb-Loop": ["1"],
    "X-Cdn-Domain": ["www.bitballoon.com"],
    "X-Country": ["US"],
    "X-Datadog-Parent-Id": ["1524629822969833135"],
    "X-Datadog-Sampling-Priority": ["1"],
    "X-Datadog-Trace-Id": ["15496261542005402192"],
    "X-Forwarded-For": ["67.180.63.210, 100.64.0.115"],
    "X-Forwarded-Proto": ["https"],
    "X-Language": ["en,en;q=0.9,fr;q=0.8"],
    "X-Nf-Client-Connection-Ip": ["67.180.63.210"],
    "X-Nf-Connection-Proto": ["https"],
    "X-Nf-Request-Id": ["01FFEGH1HV7GPWHK5HCGQ4AAAV"],
    "host": ["netlify-rust-test-function.netlify.app"]
  },
  "multiValueQueryStringParameters": {},
  "path": "/.netlify/functions/hello",
  "queryStringParameters": {},
  "rawQuery": "",
  "rawUrl": "https://netlify-rust-test-function.netlify.app/.netlify/functions/hello"
}
```

The second argument we don't cover in this course. It can contain additional data from Netlify Identity, such as the authenticated user.

We use `println` again to show when this handler is being run in the logs.

`json!` is a macro from `serde` that allows us to write out JSON directly in our Rust code. The `json!` macro returns a `serde_json::Value` which is what we need to return from our handler.

```rust
json!({"body": "Boop!"})
```

The json that we return includes the body, headers, and more. All fields are optional. This is how we'd set a few of those fields.

```rust
json!({
  "body": "Boop!",
  "headers": {
    "X-Something": "some value here"
  },
  "statusCode": "200"
})
```

At this point the application should build.

```rust
cargo build -p pokemon-api
```