Unit Testing a Rust Serverless Function
The source of truth for how our function behaves belongs to the production environment. Netlify helps out a lot here with the ability to deploy branches into the production environment, which means when we PR code into our serverless project it can be deployed and we can integration test or manually review before merging.

This early version of a "testing in production" workflow that matches common mainstream workflows is very powerful.

Combining it with feature flags to enable/disable code paths and other mechanisms can lead to faster and safer deploys.

Before we deploy, when it comes to Rust, there are a few tools we use every time we compile that give us confidence our code is operating as we intend it to. Type checking with `cargo check`, linting with `cargo clippy`, benchmarking with `cargo bench`, and the final compilation itself all contribute to how much confidence we can have in our deployments.

We have one more tool we can set up and use: `cargo test`.

Cargo comes with built-in support for approaches like unit and integration testing.

We can write a unit test for our serverless function that will allow us to build confidence that it handles event objects properly and gives us the return values we expect.

This will let us run our handler code locally, in CI, or elsewhere, simulating one event being processed.

```rust
mod tests {
    use super::*;

    #[tokio::test]
    async fn handler_handles() {
        let event = json!({});

        assert_eq!(
            handler(event.clone(), Context::default())
                .await
                .unwrap(),
            json!({"body": "Boop!"})
        )
    }
}
```

The `cfg` macro lets us conditionally include source code in our compilation. In this case our test code only gets included if we're running `cargo test` as the `test` condition is only true if the Rust compiler is in "test mode", which `cargo test` sets.

> There are [quite a few conditions](https://doc.rust-lang.org/reference/conditional-compilation.html) we can use in a cfg macro, such as for target operating systems, features, or architectures.

In this case we're conditionally include a submodule which we've called `tests` out of convention. The name of this submodule doesn't matter.

Rust modules don't necessarily reflect the file system and we can see that here, where `tests` is a sub-module that we define entirely in the same file as `main.rs`.

It's incredibly useful for unit tests to be able to access any items defined in the same file, so we `use super::*` to bring all of the items, such as `handler` into scope from the parent module.

`handler` is an async function, which means we need a tokio runtime. While Rust offers the `#[test]` macro to indicate tests, the tokio crate matches this with it's own async version of the macro called `tokio::test`.

The test macro is basically a flag to the test runner to say "execute this function as a test", since test functions are really no different than regular functions. The `tokio::test` macro additionally sets up a tokio runtime for us. We can have control over what kind of runtime tokio uses in our tests if we want, including running as single-threaded, multi-threaded, or even starting with time stopped.

`handler` again is `async`, so our test needs to be an async function as well. Test functions accept no arguments and return either `()` (which is the default return value for all functions) or a `Result`. This is similar to how `main` works.

Once we're in our test, we can construct an `event` to mock the first argument to our handler. This is going to be only things our `handler` function cares about for brevity, but will have to match the structure Netlify gives our function if we want the test to be useful.

```rust
let event = json!({});
```

In this case the handler doesn't actually use any of the event yet, so we'll use an empty object.

Rust provides a few useful helper macros for writing tests. One of these is `assert_eq`, which we'll use to test the result of the `handler` against a value we construct.

```rust
assert_eq!(
    handler(event.clone(), Context::default())
        .await
        .unwrap(),
    json!({"body": "Boop!"})
)
```

The actual `handler` call accepts the `event` we created and a `Context`. Luckily for us `Context` implements the `Default` trait, which means we can use that to construct a valid value.

We have to await the `handler` since it's an async function, and then we `unwrap` the `Result` return value which will panic and fail the test if it is an `Err`.

On the right hand side we can construct a new `Value` that is the API response we expect.

Then running `cargo test` will find this test and run it for us.

```rust
cargo test -p pokemon-api
```

The output will look something like this

```rust
running 1 test
test tests::handler_handles ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s
```