From URL paths to SQL queries
With the ability to pass in arguments to our SQL query, we need a way to pass through arguments from the `event` our function receives as an argument to the SQL query. We're going to do that through the url path.

`event.path` gives us the url path for the request to our function. By default this will be something like `/.netlify/function/bulbasaur`. When we implement redirects to clean up our API routes later, that will become `/api/pokemon/bulbasaur`.

So to get the pokemon slug we should pass to the SQL query, we'll grab the last path segment in the URL path.

First we need to give a name to the event. We haven't changed the type, it is still a `serde_json::Value`.

```rust
async fn handler(
    event: Value,
    _: Context,
) -> Result<Value, Error> {
```

We can then take the `event["path"]`, which because we're using a `serde_json::Value`, `event["path"]` might not exist. We're accessing a field at runtime here. If it exists we want it as a string, not a `Value`, which returns us an `Option<&str>`.

The `Option` type includes a function called `and_then` which lets us operate on the value of an option if it is a `Some` variant, otherwise it returns `None`. So `requested_pokemon` will be an `Option` as well.

`path.split("/")` will give us an iterator over the path segments. Notably the beginning and end of the iterator can be `""` if it starts or ends with a `/`.

`last` will consume the entire iterator until it produces its last value, which in this case is the last path segment.

`last` returns an `Option`, which will be merged with the `Option` from `as_str`. This means that `requested_pokemon` will only be `Option<&str>` not `Option<Option<&str>>`.

```rust
let requested_pokemon = event["path"]
    .as_str()
    .and_then(|path| path.split("/").last());
```

We can `match` on `requested_pokemon` to handle any errors. `Some("")` allows us to match on potentially empty values, for example when someone sends a request with a trailing slash, or if the final path segment is an empty string.

`None` is actually a hard error for us. It means there wasn't a `path` in the `event`, which is a broken expectation for us. We always expect that to exist and if it doesn't, that's a server error.

Finally we have the success case, where a `pokemon_name` was successfully retrieved from the path. This code is the same code we had before for our `handler`, with the addition of using `pokemon_name` instead of a hardcoded string.

```rust
match requested_pokemon {
    Some("") => todo!(),
    None => todo!(),
    Some(pokemon_name) => {
        let pool = MySqlPoolOptions::new()
            .max_connections(5)
            .connect(&database_url)
            .await?;
        let result = sqlx::query_as!(
            PokemonHp,
            r#"SELECT name, hp FROM pokemon WHERE slug = ?"#,
            pokemon_name
        )
        .fetch_one(&pool)
        .await?;

        Ok(json!({
            "body": serde_json::to_string(&result)?
        }))
    }
}
```

We need to update our test to account for the new logic. Our fake `event` will have a `path` field added to it that will have a Pokemon slug at the end of the segments.

This will now fetch `bulbasaur` from the database.

```rust
async fn handler_handles() {
    let event = json!({
        "path": "/api/pokemon/bulbasaur"
    });

    assert_eq!(
        handler(event.clone(), Context::default())
            .await
            .unwrap(),
        json!({
            "body": serde_json::to_string(
                &PokemonHp{
                    name: String::from("Bulbasaur"),
                    hp: 45
                },
            ).unwrap()
        })
    )
}
```

We can also ensure that the other paths can be hit when we receive invalid input. We'll write two new tests: `handler_handles_empty_pokemon` and `handler_panics_on_no_path`.

Both tests will panic because of the `todo` macro, so we can use the `should_panic` attribute macro to test against the expected panic message.

The first test will exercise the `400` codepath, which happens when the path doesn't include a Pokemon name. We could return an empty response as if we hadn't found a Pokemon, but it's most likely that an empty path segment is a client mistake, having passed in an empty string or misconstructed the path.

```rust
async fn handler_handles_empty_pokemon() {
    let event = json!({
        "path": "/api/pokemon//"
    });
    handler(event.clone(), Context::default())
        .await
        .unwrap();
}
```

The second test will cover what happens if we don't have an `event` `path` at all, which is something we really need to work correctly. This will very likely result in a 500 because we're accessing the function through Netlify's URL schemes and thus should never not have an event path.

```rust
async fn handler_panics_on_no_path() {
    let event = json!({});
    handler(event.clone(), Context::default())
        .await
        .unwrap();
}
```

Running cargo test with the database url will now pass.

```rust
DATABASE_URL=mysql://127.0.0.1 cargo test
```