Deploy your first Rust Serverless Function
To deploy to Netlify Functions we'll need a `netlify.toml`.

We'll set our functions directory to `functions/`. This is where we'll put our binaries after we build them to be deployed as serverless functions.

```rust
[functions]
directory = "functions/"

[build]
command = """
mkdir functions &&
cargo build --release --bin pokemon-api &&
cp target/release/pokemon-api functions/
"""
publish = "crates"
```

The build command we'll set using a multiline toml string. Aside from `cargo build` the important pieces of the build command are that we're using the `--release` flag and that we're only building the `pokemon-api` binary.

When we `cargo build --release` cargo places the built binaries in `target/release/`, so we can copy the file into the functions directory.

We also need to create the `functions` directory with `mkdir` so we can copy the binary into it.

Finally I've chosen to use the `crates` directory, which is all of the source code, as the `publish` directory. This will serve up all of our code publicly, so you might want to choose a different directory.

We'll also need  a `rust-toolchain.toml` file in the root of our project to tell Netlify's build system which version of Rust we want to use. In this case we specify the current `stable` release of Rust, which can change over time.

```rust
[toolchain]
channel = "stable"
```

We can push this to GitHub and create a new Netlify site from git.

Netlify has built-in support for Rust builds. However as of the recording of this course it is still experimental so you'll want to add the following environment variable to the deploy settings in your site as well. Set `NETLIFY_EXPERIMENTAL_BUILD_RUST_SOURCE` to `true`.

The site should build and deploy your new serverless Rust function.